refactor: move ArraySummary components to subfolder
