+++ app/components/arraySummary/ArraySummary.tsx
@@ -5,9 +5,9 @@ import type { ArrayConfig } from "~/services/array";
 import type { Translations } from "~/services/cms/index.server";
 import { getTranslationByKey } from "~/util/getTranslationByKey";
 import ArraySummaryItemButton from "./ArraySummaryItemButton";
-import Button from "./Button";
-import Heading from "./Heading";
-import RichText from "./RichText";
+import Button from "../Button";
+import Heading from "../Heading";
+import RichText from "../RichText";
 
 type ArraySummaryProps = {
   readonly category: string;
